power scenario
