+++ integration_tests/src/tests/mod.rs
@@ -20,7 +20,8 @@ mod market_miner_withdrawal_test;
 pub use market_miner_withdrawal_test::*;
 mod multisig_test;
 pub use multisig_test::*;
-// mod power_scenario_tests;
+mod power_scenario_tests;
+pub use power_scenario_tests::*;
 // mod publish_deals_test;
 // mod replica_update_test;
 // mod terminate_test;
